feat: renaming api folder to gptube-api because of troubles with naming files to vercel
