+++ gptube/gptube-api/generated-api/.openapi-generator/VERSION
