+++ gptube/gptube-api/generated-api/models/HandlersHelloApiMessage.ts
