+++ gptube/gptube-api/generated-api/models/ModelsBertAIResults.ts
