+++ gptube/package.json
@@ -7,7 +7,7 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "generate-api": "rimraf api/generated-api && openapi-generator-cli generate"
+    "generate-api": "rimraf gptube-api/generated-api && openapi-generator-cli generate"
   },
   "dependencies": {
     "@ant-design/icons": "^5.0.1",
